+++ jolt/jolt-core/src/jolt/subtable/left_is_zero.rs
@@ -18,7 +18,7 @@ impl<F: JoltField> LeftIsZeroSubtable<F> {
 }
 
 impl<F: JoltField> LassoSubtable<F> for LeftIsZeroSubtable<F> {
-    fn materialize(&self, M: usize) -> Vec<u32> {
+    fn materialize(&self, M: usize) -> Vec<u64> {
         // table[x | y] = (x == 0)
         let mut entries = vec![0; M];
